# apps/web/src/lib/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RouteHub API",
    "version": "1.0.0",
    "description": "API documentation for RouteHub — fleet management, payments, documents, and inspections."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://geosafety.routehub.ge",
      "description": "Production (master)"
    },
    {
      "url": "https://team2.routehub.ge",
      "description": "Team 2"
    },
    {
      "url": "https://team3.routehub.ge",
      "description": "Team 3"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie (set automatically after login)"
      },
      "cronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Bearer token matching CRON_SECRET env var (for scheduled jobs)"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/payments/stats": {
      "get": {
        "summary": "Get payment statistics",
        "description": "Returns aggregated payment statistics including totals, match rates, and amounts. Uses a DB-side RPC for performance with a client-side fallback.",
        "tags": ["Payments"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date for date range filter"
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date for date range filter"
          },
          {
            "name": "matchSource",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "paused", "ended", "one_time"]
            },
            "description": "Filter by contract match source"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total_transactions": {
                      "type": "integer"
                    },
                    "total_amount": {
                      "type": "number"
                    },
                    "matched_count": {
                      "type": "integer"
                    },
                    "matched_amount": {
                      "type": "number"
                    },
                    "unmatched_count": {
                      "type": "integer"
                    },
                    "unmatched_amount": {
                      "type": "number"
                    },
                    "ignored_count": {
                      "type": "integer"
                    },
                    "match_rate": {
                      "type": "integer",
                      "description": "Percentage of matched transactions (0-100)"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin or dispatcher access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/payments/contracts": {
      "get": {
        "summary": "Get contract terms grouped by tax ID",
        "description": "Fetches contract data from contract-related boards (active, one-time, paused, ended) using dynamic column discovery by name. Returns contracts keyed by tax ID, enriched with the earliest bank transaction date per tax ID. Response is cached for 5 minutes.",
        "tags": ["Payments"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Contracts grouped by tax ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contracts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "item_id": {
                              "type": "string"
                            },
                            "board_id": {
                              "type": "string"
                            },
                            "contract_source": {
                              "type": "string",
                              "enum": ["active", "one_time", "paused", "ended"]
                            },
                            "company_name": {
                              "type": "string"
                            },
                            "tax_id": {
                              "type": "string"
                            },
                            "monthly_amount": {
                              "type": "number",
                              "nullable": true
                            },
                            "frequency": {
                              "type": "string",
                              "nullable": true
                            },
                            "invoice_amount": {
                              "type": "number",
                              "nullable": true
                            },
                            "status": {
                              "type": "string",
                              "nullable": true
                            },
                            "start_date": {
                              "type": "string",
                              "nullable": true
                            },
                            "end_date": {
                              "type": "string",
                              "nullable": true
                            },
                            "payment_method": {
                              "type": "string",
                              "nullable": true
                            },
                            "first_payment_date": {
                              "type": "string",
                              "nullable": true
                            }
                          }
                        }
                      }
                    },
                    "boards_found": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "count": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin or dispatcher access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}